unify calling place; remove redundant file
