+++ tt_metal/tt-llk/tests/python_tests/helpers/llk_params.py
@@ -101,6 +101,10 @@ class MathOperation(Enum):
     # three offsets directly.
     SfpuSwiGLU = OpSpec("swiglu", MathOpType.SFPU_UNARY)
     Tanh = OpSpec("tanh", MathOpType.SFPU_UNARY)
+    # Typecast is dispatched by the (input, output) DataFormat pair rather than a
+    # single op, but it maps to SfpuType::typecast and runs through the shared
+    # unary-SFPU dispatch (see call_unary_sfpu_operation in sfpu_operations.h).
+    Typecast = OpSpec("typecast", MathOpType.SFPU_UNARY)
     Threshold = OpSpec("threshold", MathOpType.SFPU_UNARY)
     ReluMax = OpSpec(
         "relu_max", MathOpType.SFPU_UNARY
